requirements.txt
streamlit==1.30.0
pdfplumber==0.7.5
python-docx==0.8.11
deep-translator==1.10.0
pandas==2.3.3
sqlitedict==2.0.0
gunicorn==20.1.0
pytest==8.2.0
